Updated SynergyKM project to newer versions
- menucracker 2.2
- synergy 1.4.8

Synergy is just dropped in. You'll want to build synergy first using
CMake before building the SynergyKM project

+++ Components/Prefpane/Widgets/SPScreenView.mm
@@ -53,7 +53,7 @@
 
 @end
 
-static const float kMaxScreenSize = 100.0f;
+static const CGFloat kMaxScreenSize = 100.0f;
 
 @implementation SPScreenView
